# kfr_python_junior_test_openapi.json
{
    "openapi": "3.0.2",
    "info":
    {
        "title": "Kefir Python Junior Test",
        "version": "0.1.0"
    },
    "paths":
...
        "/login":
        {
            "post":
            {
                "tags":
                [
                    "auth"
                ],
                "summary": "Вход в систему",
                "description": "После успешного входа в систему необходимо установить Cookies для пользователя",
                "operationId": "login_login_post",
                "requestBody":
                {
                    "content":
                    {
                        "application/json":
                        {
                            "schema":
                            {
                                "$ref": "#/components/schemas/LoginModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses":
                {
                    "200":
                    {
                        "description": "Successful Response",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/CurrentUserResponseModel"
                                }
                            }
                        }
                    },
                    "400":
                    {
                        "description": "Bad Request",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422":
                    {
                        "description": "Validation Error",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/current":
        {
            "get":
            {
                "tags":
                [
                    "user"
                ],
                "summary": "Получение данных о текущем пользователе",
                "description": "Здесь находится вся информация, доступная пользователю о самом себе,\n        а так же информация является ли он администратором",
                "operationId": "current_user_users_current_get",
                "responses":
                {
                    "200":
                    {
                        "description": "Successful Response",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/CurrentUserResponseModel"
                                }
                            }
                        }
                    },
                    "400":
                    {
                        "description": "Bad Request",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401":
                    {
                        "description": "Unauthorized",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "title": "Response 401 Current User Users Current Get",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users":
        {
            "get":
            {
                "tags":
                [
                    "user"
                ],
                "summary": "Постраничное получение кратких данных обо всех пользователях",
                "description": "Здесь находится вся информация, доступная пользователю о других пользователях",
                "operationId": "users_users_get",
                "parameters":
                [
                    {
                        "required": true,
                        "schema":
                        {
                            "title": "Page",
                            "type": "integer"
                        },
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema":
                        {
                            "title": "Size",
                            "type": "integer"
                        },
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses":
                {
                    "200":
                    {
                        "description": "Successful Response",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/UsersListResponseModel"
                                }
                            }
                        }
                    },
                    "400":
                    {
                        "description": "Bad Request",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401":
                    {
                        "description": "Unauthorized",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "title": "Response 401 Users Users Get",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422":
                    {
                        "description": "Validation Error",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{pk}":
        {
            "patch":
            {
                "tags":
                [
                    "user"
                ],
                "summary": "Изменение данных пользователя",
                "description": "Здесь пользователь имеет возможность изменить свои данные",
                "operationId": "edit_user_users__pk__patch",
                "parameters":
                [
                    {
                        "required": true,
                        "schema":
                        {
                            "title": "Pk",
                            "type": "integer"
                        },
                        "name": "pk",
                        "in": "path"
                    }
                ],
                "requestBody":
                {
                    "content":
                    {
                        "application/json":
                        {
                            "schema":
                            {
                                "$ref": "#/components/schemas/UpdateUserModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses":
                {
                    "200":
                    {
                        "description": "Successful Response",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/UpdateUserResponseModel"
                                }
                            }
                        }
                    },
                    "400":
                    {
                        "description": "Bad Request",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401":
                    {
                        "description": "Unauthorized",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "title": "Response 401 Edit User Users  Pk  Patch",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404":
                    {
                        "description": "Not Found",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "title": "Response 404 Edit User Users  Pk  Patch",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422":
                    {
                        "description": "Validation Error",
                        "content":
                        {
                            "application/json":
                            {
                                "schema":
                                {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components":
...
        "schemas":
...
            "CurrentUserResponseModel":
            {
                "title": "CurrentUserResponseModel",
                "required":
                [
                    "first_name",
                    "last_name",
                    "other_name",
                    "email",
                    "phone",
                    "birthday",
                    "is_admin"
                ],
                "type": "object",
                "properties":
                {
                    "first_name":
                    {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name":
                    {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "other_name":
                    {
                        "title": "Other Name",
                        "type": "string"
                    },
                    "email":
                    {
                        "title": "Email",
                        "type": "string"
                    },
                    "phone":
                    {
                        "title": "Phone",
                        "type": "string"
                    },
                    "birthday":
                    {
                        "title": "Birthday",
                        "type": "string",
                        "format": "date"
                    },
                    "is_admin":
                    {
                        "title": "Is Admin",
                        "type": "boolean"
                    }
                }
            },
            "ErrorResponseModel":
            {
                "title": "ErrorResponseModel",
                "required":
                [
                    "code",
                    "message"
                ],
                "type": "object",
                "properties":
                {
                    "code":
                    {
                        "title": "Code",
                        "type": "integer"
                    },
                    "message":
                    {
                        "title": "Message",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError":
            {
                "title": "HTTPValidationError",
                "type": "object",
                "properties":
                {
                    "detail":
                    {
                        "title": "Detail",
                        "type": "array",
                        "items":
                        {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "LoginModel":
            {
                "title": "LoginModel",
                "required":
                [
                    "login",
                    "password"
                ],
                "type": "object",
                "properties":
                {
                    "login":
                    {
                        "title": "Login",
                        "type": "string"
                    },
                    "password":
                    {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
...
            "UpdateUserModel":
            {
                "title": "UpdateUserModel",
                "type": "object",
                "properties":
                {
                    "first_name":
                    {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name":
                    {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "other_name":
                    {
                        "title": "Other Name",
                        "type": "string"
                    },
                    "email":
                    {
                        "title": "Email",
                        "type": "string"
                    },
                    "phone":
                    {
                        "title": "Phone",
                        "type": "string"
                    },
                    "birthday":
                    {
                        "title": "Birthday",
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "UpdateUserResponseModel":
            {
                "title": "UpdateUserResponseModel",
                "required":
                [
                    "id",
                    "first_name",
                    "last_name",
                    "other_name",
                    "email",
                    "phone",
                    "birthday"
                ],
                "type": "object",
                "properties":
                {
                    "id":
                    {
                        "title": "Id",
                        "type": "integer"
                    },
                    "first_name":
                    {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name":
                    {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "other_name":
                    {
                        "title": "Other Name",
                        "type": "string"
                    },
                    "email":
                    {
                        "title": "Email",
                        "type": "string"
                    },
                    "phone":
                    {
                        "title": "Phone",
                        "type": "string"
                    },
                    "birthday":
                    {
                        "title": "Birthday",
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "UsersListMetaDataModel":
            {
                "title": "UsersListMetaDataModel",
                "required":
                [
                    "pagination"
                ],
                "type": "object",
                "properties":
                {
                    "pagination":
                    {
                        "$ref": "#/components/schemas/PaginatedMetaDataModel"
                    }
                }
            },
            "UsersListResponseModel":
            {
                "title": "UsersListResponseModel",
                "required":
                [
                    "data",
                    "meta"
                ],
                "type": "object",
                "properties":
                {
                    "data":
                    {
                        "title": "Data",
                        "type": "array",
                        "items":
                        {
                            "$ref": "#/components/schemas/UsersListElementModel"
                        }
                    },
                    "meta":
                    {
                        "$ref": "#/components/schemas/UsersListMetaDataModel"
                    }
                }
            },
...
        }
    }
}